Fix bug when deleting higher level asset fields
If a higher level asset sets a field but then the field is deleted, it
shouldn't do anything. Previously it would save a new blank entry at the
current level. It doesn't affect the parent entry.

This Previously wasn't a problem, but it is intended to only save the
parent field as a placeholder. If this is the case, then forms with
blank fields will be submitted regularly but this does not mean replace
the field with a blank.

+++ app/models/concerns/has_asset_record.rb
@@ -29,9 +29,9 @@ module HasAssetRecord
     asset_record.each do |field|
       updated_value = definition_hash[field.definition.id.to_s.to_sym]
       next if field.value == updated_value
-      if field.asset == self && (updated_value.nil? || updated_value.empty?)
+      if updated_value.nil? || updated_value.empty?
         # Delete an existing field
-        field.destroy!
+        field.destroy! if field.asset == self
       elsif field.asset == self
         # When updating a field associated with the asset
         field.value = updated_value
